+++ packages/expo-modules-core/android/src/main/cpp/decorators/JSDecoratorsBridgingObject.h
@@ -13,6 +13,7 @@
 
 #include "JSFunctionsDecorator.h"
 #include "JSPropertiesDecorator.h"
+#include "JSLazyPropertiesDecorator.h"
 #include "JSConstantsDecorator.h"
 #include "JSObjectDecorator.h"
 #include "JSClassesDecorator.h"
@@ -39,6 +40,11 @@ public:
     jboolean setterTakesOwner,
     jni::alias_ref<jni::JArrayClass<ExpectedType>> setterExpectedArgsTypes,
     jni::alias_ref<JNIFunctionBody::javaobject> setter
+  );
+
+  void registerLazyProperty(
+    jni::alias_ref<jstring> name,
+    jni::alias_ref<JNINoArgsFunctionBody::javaobject> getter
   );
 
   void registerSyncFunction(
@@ -86,6 +92,7 @@ private:
   std::unique_ptr<JSFunctionsDecorator> functionDecorator;
   std::unique_ptr<JSConstantsDecorator> constantsDecorator;
   std::unique_ptr<JSPropertiesDecorator> propertiesDecorator;
+  std::unique_ptr<JSLazyPropertiesDecorator> lazyPropertiesDecorator;
   std::unique_ptr<JSObjectDecorator> objectDecorator;
   std::unique_ptr<JSClassesDecorator> classDecorator;
 };